refactor: showActivityAlert 메서드 extension으로 수정
